NetworkDiagrammer/requirements.txt

PyQt6==6.5.0
python-nmap==0.7.1
scapy==2.5.0
networkx==3.1
graphviz==0.20.1
matplotlib==3.7.1
